Development/custom codes2 (#2008)
* [Core] Add uint24_t (poor mans solution)

* [Core] uint24_t and int24_t as full supported types

* [Core] add new files to CMakeList.txt
