+++ Source/core/Number.h
@@ -25,8 +25,6 @@
 #include "TextFragment.h"
 #include "TypeTraits.h"
 
-#include <limits>
-
 namespace Thunder {
 namespace Core {
     extern "C" {
@@ -1051,94 +1049,6 @@ namespace Core {
         uint8_t _size;
     };
 
-    class SInt24 {
-    public:
-        static constexpr uint8_t SizeOf = 3;
-        static constexpr uint32_t Max = 0x7FFFFF;
-        static constexpr int32_t Min = 0xFF800000;
-
-        using InternalType = int32_t;
-
-        SInt24()
-            : _value(0)
-        {
-        }
-        SInt24(const int32_t value)
-            : _value(value | (value & 0x800000 ? 0xFF000000 : 0))
-        {
-            ASSERT(((static_cast<uint32_t>(value) >> 24) == 0) || ((static_cast<uint32_t>(value) >> 24) == 0xFF));
-        }
-        SInt24(const SInt24&) = default;
-        SInt24(SInt24&&) = default;
-        ~SInt24() = default;
-
-    public:
-        SInt24& operator=(const SInt24& value) = default;
-        SInt24& operator=(SInt24&& value) = default;
-        SInt24& operator=(const int32_t value)
-        {
-            ASSERT(((static_cast<uint32_t>(value) >> 24) == 0) || ((static_cast<uint32_t>(value) >> 24) == 0xFF));
-            _value = (value | (value & 0x800000 ? 0xFF000000 : 0));
-            return (*this);
-        }
-        operator int32_t() const
-        {
-            return (_value);
-        }
-
-        int32_t AsSInt24() const
-        {
-            return (_value & 0xFFFFFF);
-        }
-
-    private:
-        int32_t _value;
-    };
-
-    class UInt24 {
-    public:
-        static constexpr uint8_t SizeOf = 3;
-        static constexpr uint32_t Max = 0xFFFFFF;
-        static constexpr uint32_t Min = 0;
-
-        using InternalType = uint32_t;
-
-        UInt24()
-            : _value(0)
-        {
-        }
-        UInt24(const UInt24& value) = default;
-        UInt24(UInt24&& value) = default;
-        UInt24(const uint32_t value)
-            : _value(value)
-        {
-            ASSERT((value >> 24) == 0);
-        }
-        ~UInt24() = default;
-
-    public:
-        UInt24& operator=(const UInt24& copy) = default;
-        UInt24& operator=(UInt24&& move) = default;
-        UInt24& operator=(const uint32_t value)
-        {
-            ASSERT((value >> 24) == 0);
-            _value = value;
-            return (*this);
-        }
-        operator uint32_t() const
-        {
-            return (_value);
-        }
-
-        uint32_t AsUInt24() const // just to be consistent with SInt24
-        {
-            return (_value & 0xFFFFFF); // in debug assert should already have fired on assignment
-        }
-
-    private:
-        uint32_t _value;
-    };
-
     template <typename T, typename std::enable_if<std::is_scalar<T>::value, int>::type = 0>
     static constexpr uint8_t RealSize()
     {
@@ -1148,6 +1058,25 @@ namespace Core {
     static constexpr uint8_t RealSize()
     {
         return (T::SizeOf);
+    }
+
+    // -----------------------------------------------------
+    // Check for Overflowed member available on number type
+    // -----------------------------------------------------
+    IS_MEMBER_AVAILABLE(Overflowed, hasOverflowed);
+
+    template <typename TYPE>
+    inline typename Core::TypeTraits::enable_if<hasOverflowed<const TYPE, bool>::value, bool>::type
+    Overflowed(TYPE t)
+    {
+        return t.Overflowed();
+    }
+
+    template <typename TYPE>
+    inline typename Core::TypeTraits::enable_if<!hasOverflowed<const TYPE, bool>::value, bool>::type
+    Overflowed(TYPE)
+    {
+        return false;
     }
         
     template <typename NEW_TYPE, typename ORIGINAL_TYPE>
@@ -1175,107 +1104,6 @@ namespace Core {
 
 } // namespace Core
 } // namespace Thunder
-
-namespace std { // seems to be allowed/mandatory to specialize inside the std namespace
-
-template <>
-class numeric_limits<Thunder::Core::SInt24> {
-
-public:
-    static constexpr bool is_specialized = true;
-
-    static constexpr Thunder::Core::SInt24::InternalType min() noexcept { return Thunder::Core::SInt24::Min; }
-    static constexpr Thunder::Core::SInt24::InternalType max() noexcept { return Thunder::Core::SInt24::Max; }
-    static constexpr Thunder::Core::SInt24::InternalType lowest() noexcept { return min(); }
-
-    static constexpr int digits = 23;
-    static constexpr int digits10 = 6;
-    static constexpr int max_digits10 = 0;
-
-    static constexpr bool is_signed = true;
-    static constexpr bool is_integer = true;
-    static constexpr bool is_exact = true;
-    static constexpr int radix = 2;
-
-    static constexpr Thunder::Core::SInt24::InternalType epsilon() noexcept { return 0; }
-    static constexpr Thunder::Core::SInt24::InternalType round_error() noexcept { return 0; }
-
-    static constexpr int min_exponent = 0;
-    static constexpr int min_exponent10 = 0;
-    static constexpr int max_exponent = 0;
-    static constexpr int max_exponent10 = 0;
-
-    static constexpr bool has_infinity = false;
-    static constexpr bool has_quiet_NaN = false;
-    static constexpr bool has_signaling_NaN = false;
-    static constexpr std::float_denorm_style has_denorm = std::denorm_absent;
-    static constexpr bool has_denorm_loss = false;
-
-    static constexpr Thunder::Core::SInt24::InternalType infinity() noexcept { return static_cast<Thunder::Core::SInt24::InternalType>(0); }
-    static constexpr Thunder::Core::SInt24::InternalType quiet_NaN() noexcept { return static_cast<Thunder::Core::SInt24::InternalType>(0); }
-    static constexpr Thunder::Core::SInt24::InternalType signaling_NaN() noexcept { return static_cast<Thunder::Core::SInt24::InternalType>(0); }
-    static constexpr Thunder::Core::SInt24::InternalType denorm_min() noexcept { return static_cast<Thunder::Core::SInt24::InternalType>(0); }
-
-    static constexpr bool is_iec559 = false;
-    static constexpr bool is_bounded = true;
-    static constexpr bool is_modulo = false;
-
-    static constexpr bool traps = true;
-    static constexpr bool tinyness_before = false;
-    static constexpr std::float_round_style round_style = std::round_toward_zero;
-};
-
-template <>
-class numeric_limits<Thunder::Core::UInt24> {
-
-public:
-    static constexpr bool is_specialized = true;
-
-    static constexpr Thunder::Core::UInt24::InternalType min() noexcept { return Thunder::Core::UInt24::Min; }
-    static constexpr Thunder::Core::UInt24::InternalType max() noexcept { return Thunder::Core::UInt24::Max; }
-    static constexpr Thunder::Core::UInt24::InternalType lowest() noexcept { return min(); }
-
-    static constexpr int digits = 24;
-    static constexpr int digits10 = 7;
-    static constexpr int max_digits10 = 0;
-
-    static constexpr bool is_signed = false;
-    static constexpr bool is_integer = true;
-    static constexpr bool is_exact = true;
-    static constexpr int radix = 2;
-
-    static constexpr Thunder::Core::UInt24::InternalType epsilon() noexcept { return 0; }
-    static constexpr Thunder::Core::UInt24::InternalType round_error() noexcept { return 0; }
-
-    static constexpr int min_exponent = 0;
-    static constexpr int min_exponent10 = 0;
-    static constexpr int max_exponent = 0;
-    static constexpr int max_exponent10 = 0;
-
-    static constexpr bool has_infinity = false;
-    static constexpr bool has_quiet_NaN = false;
-    static constexpr bool has_signaling_NaN = false;
-    static constexpr std::float_denorm_style has_denorm = std::denorm_absent;
-    static constexpr bool has_denorm_loss = false;
-
-    static constexpr Thunder::Core::UInt24::InternalType infinity() noexcept { return static_cast<Thunder::Core::UInt24::InternalType>(0); }
-    static constexpr Thunder::Core::UInt24::InternalType quiet_NaN() noexcept { return static_cast<Thunder::Core::UInt24::InternalType>(0); }
-    static constexpr Thunder::Core::UInt24::InternalType signaling_NaN() noexcept { return static_cast<Thunder::Core::UInt24::InternalType>(0); }
-    static constexpr Thunder::Core::UInt24::InternalType denorm_min() noexcept { return static_cast<Thunder::Core::UInt24::InternalType>(0); }
-
-    static constexpr bool is_iec559 = false;
-    static constexpr bool is_bounded = true;
-    static constexpr bool is_modulo = true;
-
-    static constexpr bool traps = true;
-    static constexpr bool tinyness_before = false;
-    static constexpr std::float_round_style round_style = std::round_toward_zero;
-};
-
-} // namespace std
-
-using uint24_t = Thunder::Core::UInt24;
-using int24_t = Thunder::Core::SInt24;
 
 template <typename NEW_TYPE, typename ORIGINAL_TYPE>
 NEW_TYPE int_cast(const ORIGINAL_TYPE& input)
